backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
psycopg2-binary>=2.9.1
SQLAlchemy>=1.4.0
pandas>=1.3.0
openpyxl>=3.0.7
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dateutil>=2.8.2
